# Use JaFaMa for faster math, and optimize critical code paths