… lib/jafama-2.3.0.jar
… libsrc/jafama-2.3.0-sources.jar
… pom.xml
… src/edu/stanford/nlp/coref/statistical/SimpleLinearClassifier.java
… src/edu/stanford/nlp/ie/AbstractSequenceClassifier.java
… src/edu/stanford/nlp/ie/EmpiricalNERPrior.java
… src/edu/stanford/nlp/ie/NERGUI.java
… src/edu/stanford/nlp/ie/crf/CRFBiasedClassifier.java
… src/edu/stanford/nlp/ie/crf/CRFClassifier.java
… src/edu/stanford/nlp/ie/crf/CRFCliqueTree.java
… src/edu/stanford/nlp/ie/crf/CRFDatum.java
… src/edu/stanford/nlp/ie/crf/CRFFeatureExporter.java
… src/edu/stanford/nlp/ie/crf/CRFLogConditionalObjectiveFloatFunction.java
… src/edu/stanford/nlp/ie/crf/CRFNonLinearLogConditionalObjectiveFunction.java
… src/edu/stanford/nlp/ie/crf/FactorTable.java
… src/edu/stanford/nlp/ie/crf/FloatFactorTable.java
… src/edu/stanford/nlp/ie/crf/NERGUI.java
… src/edu/stanford/nlp/ie/crf/NonLinearCliquePotentialFunction.java
… src/edu/stanford/nlp/ie/crf/NonLinearSecondOrderCliquePotentialFunction.java
… src/edu/stanford/nlp/ie/qe/UnitPrefix.java
… src/edu/stanford/nlp/ling/tokensregex/types/ValueFunctions.java
… src/edu/stanford/nlp/loglinear/inference/CliqueTree.java
… src/edu/stanford/nlp/loglinear/inference/TableFactor.java
… src/edu/stanford/nlp/loglinear/learning/BacktrackingAdaGradOptimizer.java
… src/edu/stanford/nlp/loglinear/learning/LogLikelihoodDifferentiableFunction.java
… src/edu/stanford/nlp/math/ADMath.java
… src/edu/stanford/nlp/math/ArrayMath.java
… src/edu/stanford/nlp/math/SloppyMath.java
… src/edu/stanford/nlp/maxent/iis/LambdaSolve.java
… src/edu/stanford/nlp/naturalli/ClauseSplitterSearchProblem.java