These changes substantially cut down the processing time; by several hours when I process all of Wikipedia.
Feel free to benchmark on your own data.

The first commit uses JaFaMa instead of java.lang.Math, which is 2-3x faster for exp, log: http://blog.element84.com/improving-java-math-perf-with-jafama.html
In some places I switched back to log1p, because the runtime of log and log1p in JaFaMa are similar, and log1p offers better precision for small values of x than log(1+x).

The other patches optimize the crucial code around the Viterbi algorithm:

- HotSpot optimizes better if large functions with multiple loops are split into multiple methods (as they can be recompiled independently).
- It pays off to save repeated nested array lookups (e.g. `array[i][j]` in a loop over `j`; move `array_i = array[i]` outside of the loop and use `array_i[j]` inside).
- I also add a cache to avoid recomputing the open tags set in Ttags.

All of these may appear to be trivial changes, but once you benchmark you will see how much this improves the run time.

Processing the first 20000 articles with `tokenize,ssplit,pos`, doing some further processing such as my own lemmatization based on hunspell, and then loading them into a lucene index with the CoreNLP master branch took 08:51 minutes, and with my patches only 04:38 minutes (sloopy benchmark only). I consider this a substantial speedup, because Wikipedia is 5.3 million articles, and it still needed 19 hours to build the full text index, but it used to take almost two days...