@J38 2017-06-08
Thanks for contributing this!  I am currently working on finalizing the Stanford CoreNLP 3.8.0 release but when I finish that I will look this over!