@J38 2017-09-28
So this looks really cool.  I'm going to try to talk to Chris and see if there are any reasons we shouldn't convert to this, he will have better insight than me!  But the speed ups sound awesome!