@kno10 2017-12-08
I rebased my branch onto 0c07f3d0e1d7c0a9b2188a7ba3854245e4b2e7fa because the current git version performs much worse for me (c.f., #584)

The speedup in training I get is 895.0 to 713.4, I.e. about 20% faster training (+ less memory needed).

With these patches:
```
CRFClassifier training ... done [713.4 sec].
...
CRFClassifier tagged 51645 words in 3068 documents at 26189,15 words per second.
         Entity    P       R       F1      TP      FP      FN
            LOC    0,7901  0,6215  0,6957  734     195     447
           MISC    0,7464  0,4079  0,5275  412     140     598
            ORG    0,8253  0,5633  0,6695  699     148     542
            PER    0,7497  0,4768  0,5829  668     223     733
         Totals    0,7807  0,5200  0,6242  2513    706     2320
```
without
```
CRFClassifier training ... done [895.0 sec].
...
CRFClassifier tagged 51645 words in 3068 documents at 21626,88 words per second.
         Entity    P       R       F1      TP      FP      FN
            LOC    0,7887  0,6257  0,6978  739     198     442
           MISC    0,7408  0,4188  0,5351  423     148     587
            ORG    0,8239  0,5657  0,6708  702     150     539
            PER    0,7486  0,4825  0,5868  676     227     725
         Totals    0,7784  0,5256  0,6275  2540    723     2293
```
The slight variations in quality are probably due to using log1p instead of log(1+) in some places, etc. I'd assume that increasing QNsize back to 10 will make up for that. And with these patches - because they also reduce memory consumption - I can run QNsize 10 on connl with 2 GB limit, while I could not with the unmodified version.