@kno10 2017-12-15
Hi, I have rebased my changes on the lastest git version (as #584 is fixed).
I have also reordered them such that JaFaMa is _only_ in the last commit.
If you pick the changes except the JaFaMa change, then you will get all the other optimizations, without introducing the JaFaMa dependency (runtime 840.1).