@kno10 2018-02-21
Updated benchmarks (other settings than above):
```
3.8.0          CRFClassifier training ... done [4645.4 sec].
3.9.0-20180131 CRFClassifier training ... done [4616.0 sec].
patch-jafama   CRFClassifier training ... done [3534.3 sec].

3.8.0          CRFClassifier tagged 51645 words in 3068 documents at 15602.72 words per second.
3.9.0-20180131 CRFClassifier tagged 51645 words in 3068 documents at 17808.62 words per second.
patch-jafama   CRFClassifier tagged 51645 words in 3068 documents at 26242.38 words per second.

3.8.0          Totals	0.8386	0.6872	0.7554	3321	639	1512
3.9.0-20180131 Totals	0.8386	0.6872	0.7554	3321	639	1512
patch-jafama   Totals	0.8388	0.6872	0.7555	3321	638	1512
```
I.e. training time reduced by ~23%, or from 1 hour 17 minutes to just under 1 hour.

The gains in prediction appear larger, but that is likely misleading. It reduced the time from 2.9 to 1.9 seconds (so by 30%), but the measurement accuracy is not as good, and I/O may have an impact here. Much of the optimization focuses on reducing the training time, as I am experimenting with some larger training data sets. But maybe the optimized feature extraction and Viterbi do help.

The result quality is almost identical. Because of numerical changes, this version does 1 mistake less on this test data set, but that is not significant.