@AngledLuffa 2019-09-18
How do you feel about separating this into two pull requests, one with the new library and one with the rest of the changes?