@kno10 2019-09-18
It already is (at least mostly). JaFaMa is only added in one of the last commits, 94518ae82766ceddc0692ce852b4d85229ed7084
I began further restructuring these patches to make this the last commit in the series, but never got around to finishing and pushing that rebase. 
I just force-pushed that rebased version, as it seems to now have the JaFaMa patch last.