@kno10 2019-09-18
Breaking this completely into two does not make as much sense:
Some of the earlier patches change / improve / enhance some parts of ArrayMath, for example, where JaFaMa is now used. So they are not completely independent code-wise.
Some of the other patches in this pull (such as the 2x1 character Javadoc patch) are straightforward. "Optimize feature generation" is a separate pull request (as it changes lines that use Math).