@AngledLuffa 2019-09-20
Well, I integrated the NERFeatureFactory changes (and the tiny javadoc changes).  The rest of the changes are a lot of changes to code I haven't spent much time with, so I need to go over them in a bit of detail in order to accept them.  Thanks for the contributions!  I will continue working on this.