@AngledLuffa 2019-09-20
Left a few comments on the first actual change, the one touching a bunch of
tagger code.  Mostly seems good.  Thanks for the help!

On Wed, Sep 18, 2019 at 2:18 PM Erich Schubert <notifications@github.com>
wrote:

> Breaking this completely into two does not make as much sense:
> Some of the earlier patches change / improve / enhance some parts of
> ArrayMath, for example, where JaFaMa is now used. So they are not
> completely independent code-wise.
> Some of the other patches in this pull (such as the 2x1 character Javadoc
> patch) are straightforward. "Optimize feature generation" is a separate
> pull request (as it changes lines that use Math).
>
> —
> You are receiving this because you commented.
> Reply to this email directly, view it on GitHub
> <https://github.com/stanfordnlp/CoreNLP/pull/458?email_source=notifications&email_token=AA2AYWO6CI5IJUEVDJVY7CLQKKLKBA5CNFSM4DOSDH4KYY3PNVWWK3TUL52HS4DFVREXG43VMVBW63LNMVXHJKTDN5WW2ZLOORPWSZGOD7BPNQQ#issuecomment-532870850>,
> or mute the thread
> <https://github.com/notifications/unsubscribe-auth/AA2AYWMJA72UI36JSHRCLDLQKKLKBANCNFSM4DOSDH4A>
> .
>