@kno10 2019-09-20
Pushed a new version that also refactors exact histories in a similar way.