@AngledLuffa 2019-09-20
How do you feel about rebasing now that a few of the changes have been merged?  Also, if you could push the JaFaMa change as far down as possible, that would make the process of merging the rest of the changes easier.  Thanks for contributing.