@AngledLuffa 2019-09-20
It was easy to spot when I tested the tagger :)

On Fri, Sep 20, 2019 at 11:01 AM Erich Schubert <notifications@github.com>
wrote:

> *@kno10* commented on this pull request.
> ------------------------------
>
> In src/edu/stanford/nlp/tagger/maxent/TestSentence.java
> <https://github.com/stanfordnlp/CoreNLP/pull/458#discussion_r326741712>:
>
> >      }
>
> +    String[] arr1;
>      String word = sent.get(pos - leftWindow());
>
> well spotted. I may have introduced this just yesterday with the rebases,
> because this fragment had been independently improved (I recall having had
> a compile error because of this arr1).
>
> —
> You are receiving this because you commented.
> Reply to this email directly, view it on GitHub
> <https://github.com/stanfordnlp/CoreNLP/pull/458?email_source=notifications&email_token=AA2AYWIGGYI76CLYGMPTU3DQKUFZNA5CNFSM4DOSDH4KYY3PNVWWK3TUL52HS4DFWFIHK3DMKJSXC5LFON2FEZLWNFSXPKTDN5WW2ZLOORPWSZGOCFOR3WY#discussion_r326741712>,
> or mute the thread
> <https://github.com/notifications/unsubscribe-auth/AA2AYWOTHIUP62XAFBN7XOTQKUFZNANCNFSM4DOSDH4A>
> .
>