@kno10 2019-09-20
Rebased once more.
Github does not show the patches in the correct order. The jafama patch was the last in the series the entire time. I guess github is confused because they are not in temporal order; during rebasing I moved some of the simpler patches first, and the larger such as JaFaMa last. Right now, it is also showing your last reply inbetween of the patches...