@AngledLuffa 2019-09-20
Would you check that everything has been merged from "Optimize: avoid reconstructing the same set and array again and again."